Update/티켓 스케줄링 처리
