--- src/modules/batch/task.service.ts
@@ -1,12 +1,0 @@
-import { Injectable, Logger } from '@nestjs/common';
-import { Cron } from '@nestjs/schedule';
-
-@Injectable()
-export class TaskService {
-  private readonly logger = new Logger(TaskService.name);
-
-  @Cron('* * * * * *', { name: 'cronTask' })
-  handleCron() {
-    this.logger.log('Task Called!');
-  }
-}
