+++ src/modules/task/task.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 import { ScheduleModule } from '@nestjs/schedule';
 import { TaskService } from './task.service';
 import { BatchController } from './batch.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Ticket } from 'src/entities/shows/ticket.entity';
 
 @Module({
-  imports: [ScheduleModule.forRoot()],
+  imports: [ScheduleModule.forRoot(), TypeOrmModule.forFeature([Ticket])],
   providers: [TaskService],
   controllers: [BatchController],
 })
-export class BatchModule {}
+export class TaskModule {}
